% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keboola.r.docker.application.R
\docType{class}
\name{DockerApplication-class}
\alias{DockerApplication}
\alias{DockerApplication-class}
\title{Application which serves as a backend for component which runs
 inside Docker with interface to docker-bundle.}
\description{
Application which serves as a backend for component which runs
 inside Docker with interface to docker-bundle.
}

