\name{keboola.r.docker.application-package}
\alias{keboola.r.docker.application-package}
\alias{keboola.r.docker.application}
\docType{package}
\title{
Keboola R Application for Docker
}
\description{
Base application class, this package serves only as an ancestor for other applications running in Docker, it does nothing.
}
\details{
\tabular{ll}{
Package: \tab keboola.r.docker.application\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-04\cr
License: \tab What license is it under?\cr
}
}
\author{
Maintainer: Ondrej Popelka <ondrej.popelka@keboola.com>
}
\keyword{ programming }
\examples{
# Subclass the application
}
